requirements.txt
pybit==2.4.0
opencv-python==4.6.0.66
pytesseract==0.3.9
python-telegram-bot==13.11

